Reorganise files in directory
