+++ src/main/java/seedu/duke/controllers/ModulePlannerController.java
@@ -1,7 +1,7 @@
 package seedu.duke.controllers;
 
 import seedu.duke.CompletePreqs;
-import seedu.duke.ModuleList;
+import seedu.duke.models.ModuleList;
 import seedu.duke.models.Major;
 import seedu.duke.models.Student;
 import seedu.duke.views.CommandLineView;
