+++ src/test/java/seedu/duke/ModuleListTest.java
@@ -1,6 +1,7 @@
 package seedu.duke;
 
 import org.junit.jupiter.api.Test;
+import seedu.duke.models.ModuleList;
 
 import java.io.InvalidObjectException;
